work in progress in Fruchterman-Reingold
git-svn-id: https://svn.code.sf.net/p/socnetv/code@162 fc2d5e9b-b3ce-4657-a66d-b27cdf764933

+++ trunk/src/vertex.cpp
@@ -195,22 +195,22 @@ QString Vertex::color(){
 }
 
 
-void Vertex::setX(int  x){
-	qDebug("Vertex setX with %i",x);
+void Vertex::setX(float  x){
+	qDebug("Vertex setX with %f",x);
 	m_x=x;
 }
 
-int Vertex::x(){
+float Vertex::x(){
 	return m_x;	
 }
 
 
-void Vertex::setY(int y){
-	qDebug("Vertex setY with %i", y);
+void Vertex::setY(float y){
+	qDebug("Vertex setY with %f", y);
 	m_y=y;
 }
 
-int Vertex::y(){
+float	Vertex::y(){
 	return m_y;	
 }
 
@@ -218,6 +218,7 @@ int Vertex::y(){
 QPointF Vertex::pos () { 
 	return QPointF ( x(), y() ); 
 }
+
 
 
 void Vertex::setOutLinkColor(int target, QString color){
